Firmware: "Working" PI Controller, needs to be refined

+++ Steering/Firmware/Steering_Controller/Steering_Controller/PI_Controller_WIP/PI_Controller/pi_c/pi_c/adc_pwm.h
@@ -14,6 +14,6 @@
 
 double adc_read();
 
-void analog_write(double value);
+void analog_write(int16_t value);
 
 #endif /* ADC_PWM_H_ */
